requirements.txt
flask>=3.0.0
ezdxf>=1.3.0
shapely>=2.0.0
pypdf>=4.0.0
fpdf2>=2.7.0
pillow>=10.0.0
pytesseract>=0.3.10
pymupdf>=1.24.0
python-doctr>=0.7.0

pytest

pyproj>=3.6.0
mapbox-vector-tile>=2.0.1
SQLAlchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
